+++ scripts/start-api.ps1
@@ -2,4 +2,4 @@ cd z:\repos\chronicis\src\Chronicis.Api
 clear
 
 Write-Host "Starting Chronicis API"
-dotnet watch --no-launch-profile run --no-hot-reload
+dotnet watch run --no-hot-reload