+++ scripts/start-client.ps1
@@ -1,4 +1,4 @@
 cd z:\repos\chronicis\src\Chronicis.Client.Host
 clear
 Write-Host "Starting Chronicis Client"
-dotnet watch --no-launch-profile run --no-hot-reload
+dotnet watch run --no-hot-reload